+++ src/modules/Register/Register.tsx
@@ -1,7 +1,8 @@
 import { StackNavigationProp } from '@react-navigation/stack'
 import { PurpleQuarter } from 'assets/images'
-import { Card, LargeHeader, SSO, ViewContainer } from 'components'
+import { Card, LargeHeader, ViewContainer } from 'components'
 import Button from 'components/Button'
+import SSO from 'components/SSO'
 import { AuthNavigationRoutes } from 'modules/AppNavigation/Authentication/Authentication.routes'
 import { AuthNavigatorParamsList } from 'modules/AppNavigation/Authentication/Authentication.types'
 import React from 'react'
